refactor: fixes according to mentor
